CapsLock Toggle refactoring: Replace the standard ahk auto-execute area with a method called from a static variable inside that method.
There is no need for a standard ahk auto-execute area anymore because of this method.
This method is called automatically when the static variable autoExecute is instantiated,
and since it's a static, it will only be instantiated once!

Idea provided by @nnnik#6686 on the AHK Discord Server: https://discord.gg/s3Fqygv